refactor: remove db-admin-connection.ts
BREAKING CHANGE:
- Replicate node_env check in DBAdmin.truncateDB()
- Use this ^ fn instead of truncate()
- No longer use a separate "adminKnex"

+++ packages/server-wallet/src/db-admin/db-admin.ts
@@ -27,8 +27,13 @@ export class DBAdmin {
     return this.knex.migrate.latest({directory: path.join(__dirname, '../db/migrations')});
   }
 
-  async truncateDB(): Promise<void> {
-    const tables = [SigningWallet.tableName, Channel.tableName, Nonce.tableName];
+  async truncateDB(
+    tables = [SigningWallet.tableName, Channel.tableName, Nonce.tableName]
+  ): Promise<void> {
+    // eslint-disable-next-line no-process-env
+    if (process.env.NODE_ENV !== 'development' && process.env.NODE_ENV !== 'test') {
+      throw 'Cannot truncate unless in test or development environments';
+    }
     await Promise.all(tables.map(table => this.knex.raw(`TRUNCATE TABLE ${table} CASCADE;`)));
   }
